Using `i128` as base type as opposed to `i64`
For some constants in S1AP/NGAP signed 64 was not sufficient (while
unsigned 64 was), but to make sure we can support negative numbers as
well, we are using the `i128` as base type. This is truly an overkill
for now. But When we actually determine the bitwidth, the generated code
won't be that 'wide'.

+++ src/resolver/asn/structs/types/base.rs
@@ -23,7 +23,7 @@ pub(crate) struct Asn1ResolvedInteger {
     pub(crate) bytes: u8,
     pub(crate) signed: bool,
     pub(crate) resolved_constraints: Option<Asn1ConstraintValueSet>,
-    pub(crate) named_values: Option<HashMap<String, i64>>,
+    pub(crate) named_values: Option<HashMap<String, i128>>,
 }
 
 impl Default for Asn1ResolvedInteger {
@@ -41,9 +41,9 @@ impl Default for Asn1ResolvedInteger {
 pub(crate) struct Asn1ResolvedEnumerated {
     bytes: u8,
     signed: bool,
-    values: Option<BTreeSet<i64>>,
-    excepts: Option<BTreeSet<i64>>,
-    named_values: HashMap<String, i64>,
+    values: Option<BTreeSet<i128>>,
+    excepts: Option<BTreeSet<i128>>,
+    named_values: HashMap<String, i128>,
 }
 
 impl Default for Asn1ResolvedEnumerated {
